requirements.txt
sentence-transformers
PyPDF2
faiss-cpu
google_adk
google-generativeai
python-dotenv
numpy<2
